Migrate RepositoryConfigurationFactory to spring-boot-aether
